+++ Cargo.toml
@@ -7,3 +7,7 @@ members = [
 
 [profile.coverage]
 inherits = "dev"
+
+[profile.profiling]
+inherits = "release"
+debug = true
