+++ src/alumina-boot/Cargo.toml
@@ -12,7 +12,7 @@ quote = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 proc-macro2 = "1"
-itertools = "0.12"
+itertools = "0.13"
 
 [dependencies]
 regex = "1"
@@ -27,6 +27,6 @@ thiserror = "1"
 indexmap = "2"
 once_cell = "1"
 colored = "2"
-rustc-hash = "1"
+rustc-hash = "2"
 strum = "0.26"
 strum_macros = "0.26"
