+++ src/alumina-boot/src/src/scope.rs
@@ -1,8 +1,8 @@
 use crate::ast::{Attribute, Id, ItemP, Span};
 use crate::common::{CodeDiagnostic, CodeError, FileId, HashSet, IndexMap, Marker};
 use crate::diagnostics::DiagnosticContext;
-use crate::name_resolution::path::{Path, PathSegment};
 use crate::parser::ParseCtx;
+use crate::src::path::{Path, PathSegment};
 
 use alumina_boot_macros::AstSerializable;
 use indexmap::map::Entry;
